Upgrade to Scala 2.13.2 for all the examples.
The `book` project stays on 2.12.11, since Scalatex does not
support Scala 2.13.x.

We remove the `acyclic` plugin from the examples and only keep it
in the `book`, since it does not support Scala 2.13.x, and was not
used in the examples anyway.

We also upgrade to scala-async 0.10.0, which is the first version
that supports Scala 2.13.x.