Big revision of the whole book for Scala.js 1.1.0.
The book had stayed stuck in around 2016, using very old versions
of Scala, Scala.js, and libraries. This commit is a full revision
of the book to use all the latest versions, along with the right
explanations everywhere.

One imporant note is that the workbench-example-app in Getting
Started hasn't been updated upstream, but the new text already
assumes that it has. So this commit should not be merged yet.